Added decorators to be used for logging of class methods.
The core.loggers module contains the following decorators

core.loggers.log(level)
core.loggers.debug
core.loggers.info
core.loggers.warning
core.loggers.critical
core.loggers.error

which should be used to decorate class methods and automatically
log when a class is called and the arguments used to call them.
Different decorators can (and should) be used on class methods
with different importance. For example, a class method that is
only used when an error occurs should be decorating like so

    @error
    def phase_error(self, exception):
    	return exception

will generate an error in the logs when the method is called,
as well as actually calling the function (and raising the error
in this case). All class methods _should_ be logged but don't
necessarily have to be, if you don't want it logged simply don't
decorate it. The default however should be to use the 'debug'
log decorator, for example

    @debug
    def __init__(self, *args):
    	pass

which will generate debug values when the class constructor is
called.