Expanded BEE2CorrelationProvider to actually do something useful.
Also made the get/set_integration_time client methods aware of what
sync is being used to clock integration times. This should make setting
the integration times transparent to the user.

The serve_sma script now properly sets the right baselines to be
included for the server (it was setting them in 'input' notation
but now correctly uses 'antenna' notation).