BasicNetworkClient now uses an acknowledge transaction tuple to verify requests.
Subclasses of BasicNetworkClient can now either override the request method
or redefine the instances 'ack_trans' member to define a transaction protocol, e.g.
the iBOB client sets ack_trans='\x06\n','\rno match:\x06\n\r' meaning that the client
sends a ctrl-f and carriage return and expects back a "no match" from the server.

Note: this is similar to using the 'telnetlib' and the 'read_until' function but
I've verified that this version is roughly 15% faster.