Added a 'get_correlation' method to the SMA server and client.
Basically this is a TCP bridge to a UDP correlator client running on the
same machine as the server. Having a direct UDP connection to the correlator
is preferable but from behind a firewall is sometimes impossible. This method
can now be used in place of the client's method by the same name and, though
it will be much slower, can be securely and easily tunnelled past a firewall.