Added the PhaseTracker class for correcting phase offsets.
Currently this class implements a proportional-only controller
that uses a moving average of the phase history as the error.
Also it checks two thresholds before feeding back a computed
phase correction:

1) the RMS of the phase history must be smaller than a certain,
arbitrary value to ensure that there is actually a fringe and that
a phase is not applied to random noise, and

2) the computed phase correction must be larger than a certain,
somewhat arbitrary value to limit oscillations.

The mechanism also waits 3 integrations after any given phase
correction to let the phase settle down, this might be increased to
a full width of the moving average in the future.

This appears to work for the most part, abrupt changes in phase
will increase the phase history RMS and thus prevent the phase tracker
from updating the phase until the history settles down.