The previous commit was not complete, now it is.
The user can play with the phase and delay offsets but the delay tracking
and fringe stopping loop phase control over the baseline delays and phases.
It applies the offsets after calculating the necessary corrections, this
means that any offsets applied must wait for the next iteration of that loop
before being applied.